fix order formatting service
